create initial data not only when app registers an account but also logs in
