+++ RealmTasks Android/app/src/main/java/io/realm/realmtasks/RegisterActivity.java
@@ -32,10 +32,7 @@ import android.widget.Toast;
 
 import io.realm.Credentials;
 import io.realm.ObjectServerError;
-import io.realm.Realm;
 import io.realm.User;
-import io.realm.realmtasks.model.TaskList;
-import io.realm.realmtasks.model.TaskListList;
 
 import static android.text.TextUtils.isEmpty;
 import static io.realm.realmtasks.RealmTasksApplication.AUTH_URL;
@@ -142,21 +139,6 @@ public class RegisterActivity extends AppCompatActivity {
 
     private void registrationComplete(User user) {
         UserManager.setActiveUser(user);
-        Realm realm = Realm.getDefaultInstance();
-        realm.executeTransaction(new Realm.Transaction() {
-            @Override
-            public void execute(Realm realm) {
-                if (realm.isEmpty()) {
-                    final TaskListList taskListList = realm.createObject(TaskListList.class, 0);
-                    final TaskList taskList = new TaskList();
-                    taskList.setId(RealmTasksApplication.DEFAULT_LIST_ID);
-                    taskList.setText(RealmTasksApplication.DEFAULT_LIST_NAME);
-                    taskListList.getItems().add(taskList);
-                }
-            }
-        });
-        realm.close();
-
         Intent intent = new Intent(this, SignInActivity.class);
         intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
         startActivity(intent);
